# openApi-docs.json
{"openapi":"3.1.0","info":{"title":"E-commerce API","description":"REST API for E-commerce backend","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Test","description":"Simple test endpoints"},{"name":"Authentication","description":"Authentication APIs"}],"paths":{"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Authenticate user and return JWT access token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResultJwtResponseBody"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}},"401":{"description":"Wrong username or password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}}}}},"/api/test2":{"get":{"tags":["Test"],"summary":"Test endpoint 2","operationId":"test2","responses":{"401":{"description":"Unauthenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}},"403":{"description":"Access denied","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}},"200":{"description":"OK","content":{"text/plain":{}}}},"security":[{"bearerAuth":[]}]}},"/api/test1":{"get":{"tags":["Test"],"summary":"Test endpoint 1","operationId":"test1","responses":{"401":{"description":"Unauthenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}},"403":{"description":"Access denied","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResult"}}}},"200":{"description":"OK","content":{"text/plain":{}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"LoginRequest":{"type":"object","description":"Login credentials","properties":{"username":{"type":"string","description":"username","example":"Admin","minLength":1},"password":{"type":"string","description":"password","example":123456,"minLength":1}},"required":["password","username"]},"ApiResult":{"type":"object","description":"Standard API response wrapper","properties":{"success":{"type":"boolean","description":"Request succeeded or not","example":true},"message":{"type":"string","description":"Human-readable message","example":"Success"},"data":{"description":"Payload, present on success"},"errorCode":{"type":"string","description":"Error code, present on failure","enum":["VALIDATION","NOT_FOUND","UNAUTHORIZED","FORBIDDEN","CONFLICT","INTERNAL_ERROR"],"example":"NOT_FOUND"},"errors":{"type":"array","description":"Field-level validation errors","items":{"$ref":"#/components/schemas/FieldErrorDetail"}},"timestamp":{"type":"string","format":"date-time","description":"Response time (ISO-8601)","example":"2026-09-28T12:34:56Z"}},"required":["success","timestamp"]},"FieldErrorDetail":{"type":"object","description":"Validation error for one field","properties":{"field":{"type":"string","example":"username"},"message":{"type":"string","example":"Username is required"}},"required":["field","message"]},"ApiResultJwtResponseBody":{"type":"object","description":"Standard API response wrapper","properties":{"success":{"type":"boolean","description":"Request succeeded or not","example":true},"message":{"type":"string","description":"Human-readable message","example":"Success"},"data":{"$ref":"#/components/schemas/JwtResponseBody","description":"Payload, present on success"},"errorCode":{"type":"string","description":"Error code, present on failure","enum":["VALIDATION","NOT_FOUND","UNAUTHORIZED","FORBIDDEN","CONFLICT","INTERNAL_ERROR"],"example":"NOT_FOUND"},"errors":{"type":"array","description":"Field-level validation errors","items":{"$ref":"#/components/schemas/FieldErrorDetail"}},"timestamp":{"type":"string","format":"date-time","description":"Response time (ISO-8601)","example":"2026-09-28T12:34:56Z"}},"required":["success","timestamp"]},"JwtResponseBody":{"type":"object","description":"Login result","properties":{"accessToken":{"type":"string","description":"Jwt access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"refreshToken":{"type":"string","description":"not return in response (refresh token be handled seperately)","readOnly":true},"id":{"type":"integer","format":"int64","example":1},"username":{"type":"string","example":"Admin"},"email":{"type":"string","example":"admin@test.com"},"roles":{"type":"array","example":["Admin Role"],"items":{"type":"string"},"uniqueItems":true}},"required":["accessToken","email","id","roles","username"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}